add:removeUserById service
