+++ src/services/index.ts
@@ -1,12 +1,12 @@
-import { PrismaClient } from "@prisma/client";
-const prisma = new PrismaClient();
+// import { PrismaClient } from "@prisma/client";
+// const prisma = new PrismaClient();
 
-export const getUserList = async () => {
-  try {
-    const users = await prisma.user.findMany();
-    return users;
-  } catch (error) {
-    console.log(error);
-    return [];
-  }
-};
+// export const getUserList = async () => {
+//   try {
+//     const users = await prisma.user.findMany();
+//     return users;
+//   } catch (error) {
+//     console.log(error);
+//     return [];
+//   }
+// };
