+++ src/services/users.ts
@@ -1,0 +1,19 @@
+import { PrismaClient } from "@prisma/client";
+const prisma = new PrismaClient();
+
+export const getUserList = async () => {
+  try {
+    const users = await prisma.user.findMany();
+    return users;
+  } catch (error) {
+    throw error;
+  }
+};
+
+export const removeUserById = async (userId: string) => {
+  try {
+    await prisma.user.delete({ where: { id: userId } });
+  } catch (error) {
+    throw error;
+  }
+};
